grid_layout: drop useless double casts tripping -Werror=useless-cast
Two spots in grid_layout::arrange compute `spacing * num_gaps` where
spacing is already `double` after the Phase 2 spacing refactor.
The `static_cast<double>(...)` wrappers around those were no-ops but
went unnoticed in onyx_ui's own build. Consumers like warlords that
enable -Wuseless-cast with -Werror rejected the headers.

Now the expressions just use the native double arithmetic result.
No behaviour change; 1776 tests still pass.

+++ include/onyxui/layout/grid_layout.hh
@@ -873,7 +873,7 @@ namespace onyxui {
 
             // Calculate available width minus spacing
             int const num_gaps = std::max(0, static_cast<int>(m_column_widths.size()) - 1);
-            double const total_spacing = static_cast<double>(m_column_spacing * num_gaps);
+            double const total_spacing = m_column_spacing * num_gaps;
             double const available_for_columns = std::max(0.0, available_width - total_spacing);
 
             // Scale each column proportionally
@@ -897,7 +897,7 @@ namespace onyxui {
 
             // Calculate available height minus spacing
             int const num_gaps = std::max(0, static_cast<int>(m_row_heights.size()) - 1);
-            double const total_spacing = static_cast<double>(m_row_spacing * num_gaps);
+            double const total_spacing = m_row_spacing * num_gaps;
             double const available_for_rows = std::max(0.0, available_height - total_spacing);
 
             // Scale each row proportionally
